testing: add stack trace proposal and split out test results view
The test results view was 2k+ lines, time to split it up a bit.
